Moved methods from Interactor into storage_access_handler
